Refactorización del código
