+++ src/Tank/torreta.js
@@ -1,5 +1,5 @@
 import * as THREE from 'three';
-import createSphereGeometry from './Geometries/sphere';
+import createSphereGeometry from '../Geometries/sphere';
 import { DDSLoader } from 'three/examples/jsm/loaders/DDSLoader.js';
 
 const createTurret = () => {
